Completed functions

+++ include/image.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stddef.h>
 #include <stdint.h>
 
 #define WIDTH_HEIGH_SCALE_FACTOR (2)
@@ -7,11 +8,11 @@
 
 typedef struct
 {
-    int width, height, max_pixel_value;
-    int *pixels;
+    size_t width, height, max_pixel_value;
+    size_t *pixels;
 } GrayScaleImage;
 
-GrayScaleImage *create_grayscale_image(int width, int height, int max_pixel_value);
+GrayScaleImage *create_grayscale_image(size_t width, size_t height, size_t max_pixel_value);
 void destroy_grayscale_image(GrayScaleImage *image);
 
-GrayScaleImage *scale_grayscale_image_average(const GrayScaleImage *restrict src, const int target_width);
+GrayScaleImage *scale_grayscale_image_average(const GrayScaleImage *restrict src, const size_t target_width);
